Add files via upload
The function `` multi_UD_Sig_jug'' is the BFCNN with a batch size set of 2. The function ``multi_UD_Sign_jug_changep'' is the BFCNN with a batch size of 4. The function `` multi_UD_Sig_jug_feedf'' is the forward BFCNN for testing performance. Note that according to our code, it is easy to write a BFCNN function with any batch size, but to not require the computer to allocate too much extra memory at run time, we simulate it in a specific setting way. The ``Main_BFCNN'' is the main file.